:truck: ch3/speed 패키지 생성하여 이동
