+++ code/후/src/main/java/ch3/speed/SpeedContainer1.java
@@ -1,4 +1,4 @@
-package ch3;
+package ch3.speed;
 
 import java.util.*;
