+++ code/후/src/main/java/ch3/speed/SpeedContainer2.java
@@ -1,4 +1,4 @@
-package ch3;
+package ch3.speed;
 
 public class SpeedContainer2 {
